// packages/frosted-ui/src/lib/shared-component-types.ts
export type ColorScheme =
  | 'brand'
  | 'black'
  | 'white'
  | 'off-black'
  | 'dark-gray'
  | 'light-gray'
  | 'purple'
  | 'blue'
  | 'gray'
  | 'success-green'
  | 'warning-yellow'
  | 'error-red'
  | 'discord-purple'
  | 'twitter-blue'
  | 'tradingview-blue'
  | 'telegram-blue'
  | 'paypal-blue'
  | 'stripe-purple'
  | 'gold-gradient';

export type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl';

export type ButtonType = 'button' | 'submit' | 'reset';